+++ src/Android/Xamarin.Android/Samples/Scene/CreateTerrainSurfaceTilePackage/CreateTerrainSurfaceTilePackage.jpg
