+++ src/UWP/ArcGISRuntime.UWP.Viewer/Samples/Scene/CreateTerrainSurfaceTilePackage/CreateTerrainSurfaceTilePackage.xaml
@@ -1,5 +1,5 @@
 ﻿<UserControl
-    x:Class="ArcGISRuntime.UWP.Samples.CreateTerrainSurfaceFromTilePackage.CreateTerrainSurfaceFromTilePackage"
+    x:Class="ArcGISRuntime.UWP.Samples.CreateTerrainSurfaceTilePackage.CreateTerrainSurfaceTilePackage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:esriUI="using:Esri.ArcGISRuntime.UI.Controls">
