+++ src/iOS/Xamarin.iOS/Samples/Scene/CreateTerrainSurfaceTilePackage/CreateTerrainSurfaceTilePackage.cs
@@ -9,10 +9,13 @@
 
 using ArcGISRuntime.Samples.Managers;
 using Esri.ArcGISRuntime.Mapping;
-using Xamarin.Forms;
+using Esri.ArcGISRuntime.UI.Controls;
+using Foundation;
+using UIKit;
 
-namespace ArcGISRuntimeXamarin.Samples.CreateTerrainSurfaceFromTilePackage
+namespace ArcGISRuntimeXamarin.Samples.CreateTerrainSurfaceTilePackage
 {
+    [Register("CreateTerrainSurfaceTilePackage")]
     [ArcGISRuntime.Samples.Shared.Attributes.Sample(
         name: "Create terrain from local tile package",
         category: "Map",
@@ -20,18 +23,20 @@ namespace ArcGISRuntimeXamarin.Samples.CreateTerrainSurfaceFromTilePackage
         instructions: "When loaded, the sample will show a scene with a terrain surface applied. Pan and zoom to explore the scene and observe how the terrain surface allows visualizing elevation differences.",
         tags: new[] { "3D", "LERC", "elevation", "surface", "terrain", "tile cache" })]
     [ArcGISRuntime.Samples.Shared.Attributes.OfflineData("cce37043eb0440c7a5c109cf8aad5500")]
-    public partial class CreateTerrainSurfaceFromTilePackage : ContentPage
+    public class CreateTerrainSurfaceTilePackage : UIViewController
     {
-        public CreateTerrainSurfaceFromTilePackage()
+        // Hold references to UI controls.
+        private SceneView _mySceneView;
+
+        public CreateTerrainSurfaceTilePackage()
         {
-            InitializeComponent();
-            Initialize();
+            Title = "Create terrain surface from a tile package";
         }
 
         private void Initialize()
         {
             // Create the scene.
-            MySceneView.Scene = new Scene(Basemap.CreateImagery());
+            _mySceneView.Scene = new Scene(Basemap.CreateImagery());
 
             // Get the path to the elevation tile package.
             string packagePath = DataManager.GetDataFolder("cce37043eb0440c7a5c109cf8aad5500", "MontereyElevation.tpk");
@@ -47,11 +52,38 @@ namespace ArcGISRuntimeXamarin.Samples.CreateTerrainSurfaceFromTilePackage
             elevationSurface.ElevationSources.Add(elevationSource);
 
             // Add the surface to the scene.
-            MySceneView.Scene.BaseSurface = elevationSurface;
+            _mySceneView.Scene.BaseSurface = elevationSurface;
 
             // Set an initial camera viewpoint.
             Camera camera = new Camera(36.525, -121.80, 300.0, 180, 80.0, 0.0);
-            MySceneView.SetViewpointCamera(camera);
+            _mySceneView.SetViewpointCamera(camera);
+        }
+
+        public override void LoadView()
+        {
+            // Create the views.
+            View = new UIView() { BackgroundColor = UIColor.White };
+
+            _mySceneView = new SceneView();
+            _mySceneView.TranslatesAutoresizingMaskIntoConstraints = false;
+
+            // Add the views.
+            View.AddSubviews(_mySceneView);
+
+            // Lay out the views.
+            NSLayoutConstraint.ActivateConstraints(new[]
+            {
+                _mySceneView.TopAnchor.ConstraintEqualTo(View.SafeAreaLayoutGuide.TopAnchor),
+                _mySceneView.BottomAnchor.ConstraintEqualTo(View.BottomAnchor),
+                _mySceneView.LeadingAnchor.ConstraintEqualTo(View.LeadingAnchor),
+                _mySceneView.TrailingAnchor.ConstraintEqualTo(View.TrailingAnchor)
+            });
+        }
+
+        public override void ViewDidLoad()
+        {
+            base.ViewDidLoad();
+            Initialize();
         }
     }
 }
